requirements.txt
beautifulsoup4==4.12.3
duckduckgo_search==6.1.0
EbookLib==0.18
fastapi==0.111.0
gradio==4.32.1
langchain==0.2.1
langchain_community==0.2.1
langchain_core==0.2.3
langchain_experimental==0.0.59
langchain_openai==0.1.8
langchain_zhipu==4.1.5
loguru==0.7.2
modelscope==1.14.0
nltk==3.8.1
numexpr==2.10.0
numpy==1.26.4
pandas==2.2.2
pydantic==1.10.13
PyJWT==2.8.0
pymupdf==1.24.5
python-dotenv==1.0.1
rapidocr_onnxruntime==1.3.22
rapidocr_paddle==1.3.22
Requests==2.32.3
sentence_transformers==3.0.0
sse_starlette==2.1.0
torch==2.2.0+cu121
tqdm==4.66.2
unstructured==0.12.6
